toma orden_reco integer para generar ordenamiento

+++ app/MyDB.php
@@ -1905,7 +1905,7 @@ order by 1,2
                   frac cod_fraccion, radio cod_radio, seg cod_segmento, 
                   ncalle, nrocatastr nro_catast, piso, casa, dpto_habit, 
                   sector, edificio, entrada, 
-                  row_number() over ( partition by prov,dpto,frac,radio,seg order by mza,lado,orden_reco) ordenamiento, 
+                  row_number() over ( partition by prov,dpto,frac,radio,seg order by mza,lado,orden_reco::integer) ordenamiento, 
                   'C' estado, tipoviv 
                 from listados_segmentados)"
             );
